// src/telegram/types/message.d.ts
import { User, Chat } from './user';
import { Audio, Document, Animation, Game, PhotoSize, Sticker, Video, Voice, Contact, Location, Venue, Poll } from './media';
import { Keyboard } from './keyboard';

declare namespace TelegramBot {
    interface Message {
        message_id: number;
        from?: User;
        date: number;
        chat: Chat;
        forward_from?: User;
        forward_from_chat?: Chat;
        forward_from_message_id?: number;
        forward_signature?: string;
        forward_sender_name?: string;
        forward_date?: number;
        reply_to_message?: Message;
        edit_date?: number;
        media_group_id?: string;
        author_signature?: string;
        text?: string;
        entities?: MessageEntity[];
        caption_entities?: MessageEntity[];
        audio?: Audio;
        document?: Document;
        animation?: Animation;
        game?: Game;
        photo?: PhotoSize[];
        sticker?: Sticker;
        video?: Video;
        voice?: Voice;
        caption?: string;
        contact?: Contact;
        location?: Location;
        venue?: Venue;
        poll?: Poll;
        new_chat_members?: User[];
        left_chat_member?: User;
        new_chat_title?: string;
        new_chat_photo?: PhotoSize[];
        delete_chat_photo?: boolean;
        group_chat_created?: boolean;
        reply_markup?: Keyboard;
    }

    interface MessageEntity {
        type: MessageEntityType;
        offset: number;
        length: number;
        url?: string;
        user?: User;
    }

    type MessageEntityType = 'mention' | 'hashtag' | 'bot_command' | 'url' | 'email' | 'bold' | 'italic' | 'code' | 'pre' | 'text_link' | 'text_mention';
}

export = TelegramBot;